Refactored the global methods out of MaltreatmentTypesPlot into Utils.js. Repeated process for ObservedFeaturesPlot. Next up, PredictiveFeaturesPlot.
